+++ src/instructions/bulk_read.rs
@@ -22,7 +22,7 @@ where
 	pub fn bulk_read_cb<Read, F>(&mut self, reads: &[Read], mut on_response: F) -> Result<(), WriteError>
 	where
 		Read: AsRef<BulkReadData>,
-		F: FnMut(&BulkReadData, Result<Response<Vec<u8>>, ReadError>),
+		F: FnMut(&BulkReadData, Result<Response<&[u8]>, ReadError>),
 	{
 		for i in 0..reads.len() {
 			for j in i + 1..reads.len() {
@@ -55,7 +55,7 @@ where
 			});
 
 			match response {
-				Ok(response) => on_response(read, Ok(response.into())),
+				Ok(response) => on_response(read, Ok((&response).into())),
 				Err(e) => on_response(read, Err(e)),
 			}
 		}
@@ -79,11 +79,15 @@ where
 		let mut responses = Vec::with_capacity(reads.len());
 		let mut read_error = None;
 
-		self.bulk_read_cb(reads, |_read, data| {
+		self.bulk_read_cb(reads, |_read, response| {
 			if read_error.is_none() {
-				match data {
+				match response {
 					Err(e) => read_error = Some(e),
-					Ok(data) => responses.push(data),
+					Ok(response) => responses.push(Response {
+						motor_id: response.motor_id,
+						alert: response.alert,
+						data: response.data.to_owned()
+					}),
 				}
 			}
 		})?;