+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+main:
+  * Pass `Response<&[u8]>` to read callbacks instead of `Response<Vec<u8>>`.
+
 v0.6.1 - 2013-12-16:
   * Add `StatusPacket::error_number()`.
   * Add `MotorError::error_number()`.